add key track bench
