+++ simple-bench.js-deprecated
